avoid one redirect
