+++ src/pages/mod.rs
@@ -5,3 +5,4 @@ pub mod new_statement;
 pub mod options;
 pub mod statement;
 pub mod submissions;
+pub mod vote;
