+++ src/pages/vote.rs
@@ -1,0 +1,51 @@
+use crate::{auth::ensure_auth, next_statement::redirect_to_next_statement};
+
+use axum::{response::Redirect, Extension, Form};
+use serde::Deserialize;
+use sqlx::SqlitePool;
+use tower_cookies::Cookies;
+
+#[derive(Deserialize)]
+pub struct VoteForm {
+    statement_id: i64,
+    vote: i32,
+}
+
+pub async fn vote(
+    cookies: Cookies,
+    Extension(pool): Extension<SqlitePool>,
+    Form(vote): Form<VoteForm>,
+) -> Redirect {
+    let user = ensure_auth(&cookies, &pool).await;
+
+    sqlx::query!(
+        "INSERT INTO votes (statement_id, user_id, vote) VALUES (?, ?, ?) on conflict ( statement_id, user_id) do update set vote = excluded.vote",
+        vote.statement_id,
+        user.id,
+        vote.vote
+    )
+    .execute(&pool)
+    .await
+    .expect("Database problem");
+
+    sqlx::query!(
+        "INSERT INTO vote_history (user_id, statement_id, vote) VALUES (?, ?, ?)",
+        user.id,
+        vote.statement_id,
+        vote.vote
+    )
+    .execute(&pool)
+    .await
+    .expect("Database problem");
+
+    sqlx::query!(
+        "delete from queue where user_id = ? and statement_id = ?",
+        user.id,
+        vote.statement_id
+    )
+    .execute(&pool)
+    .await
+    .expect("Database problem");
+
+    redirect_to_next_statement(Some(user), Extension(pool)).await
+}
